% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.r
\name{bound}
\alias{bound}
\title{Internal helper function}
\usage{
bound(t = 1, r = 1, b = 0, l = 0)
}
\description{
Convenience function to create bounds
function copied directly from the productplots package
}
\author{
Hadley Wickham
}

