% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-mosaic.r
\name{check_aesthetics}
\alias{check_aesthetics}
\title{might need to overwrite check_aesthetics in geom - at the moment this function gets ignored}
\usage{
check_aesthetics(x, n)
}
\description{
might need to overwrite check_aesthetics in geom - at the moment this function gets ignored
}

