% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly}
\alias{fly}
\title{Flying Etiquette Survey Data}
\format{A data frame with 1040 rows and 27 variables:
\describe{
\item{ID}{Respondent ID}
\item{FlightFreq}{How often do you travel by plane?}
\item{DoYouRecline}{Do you ever recline your seat when you fly?}
\item{Height}{How tall are you?}
\item{Child18}{Do you have any children under 18?}
\item{Seats3_2Arms}{n a row of three seats, who should get to use the two arm rests?}
\item{Seats2_1Arm}{In a row of two seats, who should get to use the middle arm rest?}
\item{WhoControlsWindowShade}{Who should have control over the window shade?}
\item{RudeToMoveToUnsoldSeat}{Is it rude to move to an unsold seat on a plane?}
\item{RudeToTalkToNeighbor}{Generally speaking, is it rude to say more than a few words to the stranger sitting next to you on a plane?}
\item{SixHrFlightRudeToLeaveSeat}{On a six hour flight from NYC to LA, how many times is it acceptable to get up if you're not in an aisle seat?}
\item{RecliningObligationToBehind}{Under normal circumstances, does a person who reclines their seat during a flight have any obligation to the person sitting behind them?}
\item{RudeToRecline}{Is it rude to recline your seat on a plane?}
\item{EliminateReclining}{Given the opportunity, would you eliminate the possibility of reclining seats on planes entirely?}
\item{RudeToSwitchSeatsForFriends}{Is it rude to ask someone to switch seats with you in order to be closer to friends?}
\item{RudeToSwitchSeatsForFamily}{Is it rude to ask someone to switch seats with you in order to be closer to family?}
\item{RudeToWakeNeighborForBathroom}{Is it rude to wake a passenger up if you are trying to go to the bathroom?}
\item{RudeToWakeNeighborForWalk}{Is it rude to wake a passenger up if you are trying to walk around?}
\item{RudeToBringBaby}{In general, is it rude to bring a baby on a plane?}
\item{RudeToBringUnrulyChild}{In general, is it rude to knowingly bring unruly children on a plane?}
\item{UseElectronicsDuringTakeoff}{Have you ever used personal electronics during take off or landing in violation of a flight attendant's direction?	}
\item{HaveYouSmoked}{Have you ever smoked a cigarette in an airplane bathroom when it was against the rules?}
\item{Gender}{Gender}
\item{Age}{Age}
\item{HouseholdIncome}{Household Income}
\item{Education}{Education}
\item{Region}{Region}
}}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/flying-etiquette-survey}
}
\usage{
fly
}
\description{
Data from the results of a SurveyMonkey
survey commissioned by FiveThirtyEight for the story
41 Percent of Fliers Say It’s Rude To Recline Your Airplane Seat.
}
\keyword{datasets}
