% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spine.R, R/stat-spine.R
\name{geom_spine}
\alias{geom_spine}
\alias{stat_spine}
\title{Spine plots.}
\usage{
geom_spine(mapping = NULL, data = NULL, stat = "spine",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = FALSE, ...)

stat_spine(mapping = NULL, data = NULL, geom = "rect",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Spine plots.
}
\examples{
data(Titanic)
titanic <- as.data.frame(Titanic)
# ggplot(data=titanic) + geom_spine()
df <- read.csv("inst/mosaic-rects.csv")
ggplot() + geom_spine(aes(xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax), data=df)
}

