% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic-package.R
\docType{package}
\name{ggmosaic-package}
\alias{ggmosaic}
\alias{ggmosaic-package}
\title{ggmosaic: Mosaic Plots in the 'ggplot2' Framework}
\description{
Mosaic plots in the 'ggplot2' framework. Mosaic plot functionality is provided in a single 'ggplot2' layer by calling the geom 'mosaic'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haleyjeppson/ggmosaic}
  \item Report bugs at \url{https://github.com/haleyjeppson/ggmosaic}
}

}
\author{
\strong{Maintainer}: Haley Jeppson \email{hjeppson@iastate.edu}

Authors:
\itemize{
  \item Heike Hofmann \email{hofmann@iastate.edu}
  \item Di Cook \email{dicook@monash.edu}
}

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} [contributor]
}

}
\keyword{internal}
