% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic_app.R
\name{ggmosaic_app}
\alias{ggmosaic_app}
\title{Launch shiny app}
\usage{
ggmosaic_app(...)
}
\arguments{
\item{...}{arguments passed on.}
}
\description{
Shiny app "EDA with Mosaic Plots" for interactive exploratory model building
}
