% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.R
\name{hbar}
\alias{hbar}
\title{Horizontal bar partition: width constant, height varies.}
\usage{
hbar(data, bounds, offset = 0.02, max = NULL)
}
\arguments{
\item{data}{bounds data frame}

\item{bounds}{bounds of space to partition}

\item{offset}{space between spines}

\item{max}{maximum value}
}
\description{
Horizontal bar partition: width constant, height varies.
}
