% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.R
\name{hspine}
\alias{hspine}
\title{Horizontal spine partition: height constant, width varies.}
\usage{
hspine(data, bounds, offset = offset, max = NULL)
}
\arguments{
\item{data}{bounds data frame}

\item{bounds}{bounds of space to partition}

\item{offset}{space between spines}

\item{max}{maximum value}
}
\description{
Horizontal spine partition: height constant, width varies.
}
