% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.r
\name{margin}
\alias{margin}
\title{Internal helper function}
\usage{
margin(table, marginals = c(), conditionals = c())
}
\description{
function copied directly from the productplots package
}
\author{
Hadley Wickham
}

