% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.R
\name{mosaic}
\alias{mosaic}
\title{Template for a mosaic plot.
A mosaic plot is composed of spines in alternating directions.}
\usage{
mosaic(direction = "h")
}
\arguments{
\item{direction}{direction of first split}
}
\description{
Template for a mosaic plot.
A mosaic plot is composed of spines in alternating directions.
}
