% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-mosaic.r
\name{product}
\alias{product}
\title{Product}
\usage{
product(x, ...)
}
\arguments{
\item{x}{variable}

\item{...}{other arguments passed on}
}
\description{
XXX it might be good to make the dot, the colon and the hyphen parameters instead ... and then pass them on to expand_variable
}

