% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-mosaic.r
\name{product2}
\alias{product2}
\title{Wrapper for a list}
\usage{
product2(x, ...)
}
\description{
Wrapper for a list
}
\examples{
data(Titanic)
titanic <- as.data.frame(Titanic)
titanic$Survived <- factor(titanic$Survived, levels=c("Yes", "No"))
ggplot(data=titanic) +
  geom_mosaic(aes(weight=Freq, x=product2(Survived, Class), fill=Survived))
}

