% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rochdale}
\alias{rochdale}
\title{Rochdale data}
\format{A data frame consisting of binary variables:
\itemize{
  \item EconActive
  \item Wife.sAge
  \item HusbandEmployed Child
  \item Wife.sEduc
  \item Husband.sEduc
  \item Asian
  \item Household
}}
\usage{
rochdale
}
\description{
Data from a survey of 665 households in Rochdale discussed in Whittaker.
}
\references{
Whittaker
}
\keyword{datasets}

