% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\docType{data}
\name{scale_product}
\alias{ScaleContinuousProduct}
\alias{scale_product}
\alias{scale_x_product}
\alias{scale_y_product}
\title{Product scales for mosaic plots}
\format{An object of class \code{ScaleContinuousProduct} (inherits from \code{ScaleContinuousPosition}, \code{ScaleContinuous}, \code{Scale}, \code{ggproto}) of length 4.}
\usage{
scale_x_product(name = waiver(), breaks = product_breaks(),
  minor_breaks = NULL, labels = product_labels(), limits = NULL,
  expand = waiver(), oob = scales:::censor, na.value = NA_real_,
  trans = "identity", position = "bottom", sec.axis = waiver())

scale_y_product(name = waiver(), breaks = waiver(),
  minor_breaks = waiver(), labels = waiver(), limits = NULL,
  expand = waiver(), oob = scales:::censor, na.value = NA_real_,
  trans = "identity", position = "left", sec.axis = waiver())

ScaleContinuousProduct
}
\description{
product scales are especially introduced for use with mosaic plots: they are a hybrid of
continuous and discrete scales.
}
\keyword{datasets}

