% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\name{scale_type.productlist}
\alias{scale_type.productlist}
\title{Helper function for determining scales}
\usage{
\method{scale_type}{productlist}(x)
}
\arguments{
\item{x}{variable}
}
\value{
character string "productlist"
}
\description{
Used internally to determine class of variable x
}
