% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\name{scale_type.tbl_df}
\alias{scale_type.tbl_df}
\title{Helper function for determining scales}
\usage{
\method{scale_type}{tbl_df}(x)
}
\arguments{
\item{x}{variable}
}
\value{
character string "productlist"
}
\description{
Used internally, might not needed to be exported. HH: XXXX let's check
}
