% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-mosaic.r
\name{set.separators}
\alias{get.separators}
\alias{set.separators}
\title{Separator characters}
\usage{
set.separators(seps)

get.separators()
}
\arguments{
\item{seps}{Vector of length three. The first separator tags the value with its order, the second separator distinguishes between the key (variable name) and the value of the variable in the record.
The third separator links different values. XXX Can't use "|" for the first separator - not sure why. Might be a bug. XXX}
}
\description{
separators are a list of three special characters. These characters need to be different for each level.
}

