% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear.R
\name{shortcut_to_formula}
\alias{shortcut_to_formula}
\title{Translate shortcut strings to formulas}
\usage{
shortcut_to_formula(shortcut, vars, conds = NULL)
}
\arguments{
\item{shortcut}{Character: "independence", "saturated", or "conditional"}

\item{vars}{Character vector of margin variable names}

\item{conds}{Character vector of conditioning variable names (optional)}
}
\value{
Formula object
}
\description{
Translate shortcut strings to formulas
}
\keyword{internal}
