% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.R
\name{spine}
\alias{spine}
\title{Spine partition: divide longest dimension.}
\usage{
spine(data, bounds, offset = offset, max = NULL)
}
\arguments{
\item{data}{bounds data frame}

\item{bounds}{bounds of space to partition}

\item{offset}{space between spines}

\item{max}{maximum value}
}
\description{
Spine partition: divide longest dimension.
}
