% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.R
\name{squeeze}
\alias{squeeze}
\title{Internal helper function}
\usage{
squeeze(pieces, bounds = bound())
}
\description{
Squeeze pieces to lie within specified bounds; directly copied from package productplots
}
\author{
Hadley Wickham
}

