% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Passengers and crew on board the Titanic}
\format{A data frame with 2201 rows and 4 variables:
\describe{
\item{Class}{factor variable containing the class of a passenger (1st, 2nd, 3rd) or crew.}
\item{Sex}{Male/Female.}
\item{Age}{Child/Adult. This information is not very reliable, because it was inferred from boarding documents that did not state actual age in years.}
\item{Survived}{Yes/No.}
}}
\usage{
titanic
}
\description{
A dataset containing some demographics and survival of people on board the Titanic
}
\keyword{datasets}
