% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.R
\name{vspine}
\alias{vspine}
\title{Vertical spine partition: width constant, height varies.}
\usage{
vspine(data, bounds, offset = offset, max = NULL)
}
\arguments{
\item{data}{bounds data frame}

\item{bounds}{bounds of space to partition}

\item{offset}{space between spines}

\item{max}{maximum value}
}
\description{
Vertical spine partition: width constant, height varies.
}
