% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.r
\name{weighted.table}
\alias{weighted.table}
\title{Internal helper function}
\usage{
weighted.table(vars, wt = NULL)
}
\description{
function copied directly from the productplots package
}
\author{
Hadley Wickham
}

