% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-mosaic.r, R/stat-mosaic.r
\name{geom_mosaic}
\alias{geom_mosaic}
\alias{stat_mosaic}
\title{Mosaic plots.}
\usage{
geom_mosaic(mapping = NULL, data = NULL, stat = "mosaic",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = FALSE, ...)

stat_mosaic(mapping = NULL, data = NULL, geom = "mosaic",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Mosaic plots.
}
\examples{
data(Titanic)
titanic <- as.data.frame(Titanic)
# library(plyr)
# ggplot(data=titanic) + geom_mosaic(aes(weight=Freq))
# ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=list(Class, Survived))) # only works with modified check_aesthetics
ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=interaction(Class, Survived), group=1))
df <- read.csv("inst/mosaic-rects.csv")
ggplot() + geom_mosaic(aes(x = xmin, xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax), data=df)
}

