% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-mosaic.r, R/stat-mosaic.r
\name{geom_mosaic}
\alias{geom_mosaic}
\alias{stat_mosaic}
\title{Mosaic plots.}
\usage{
geom_mosaic(mapping = NULL, data = NULL, stat = "mosaic",
  position = "identity", na.rm = FALSE, divider = productplots::mosaic(),
  show.legend = NA, inherit.aes = FALSE, ...)

stat_mosaic(mapping = NULL, data = NULL, geom = "mosaic",
  position = "identity", na.rm = TRUE, divider = productplots::mosaic(),
  show.legend = NA, inherit.aes = TRUE, ...)
}
\description{
Mosaic plots.
}
\examples{
data(Titanic)
titanic <- as.data.frame(Titanic)
# library(plyr)
# ggplot(data=titanic) + geom_mosaic(aes(weight=Freq))
# ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=list(Class, Survived))) # only works with modified check_aesthetics
ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=product(Class, Survived), group=1))
ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=interaction(Class, Age), group=1, fill=Survived))
gg <- ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=interaction(Survived, Class), group=1, fill=Age))
gg
gg + geom_text(aes(x = (xmin+xmax)/2, y = (ymin+ymax)/2,
label=paste(paste0("Survived: ",vars1),paste0("Class: ",vars2), sep="\\n")), data=subset(ggplot_build(gg)$data[[1]], level==2))
# doing the right thing, but we need labelling to make it less confusing
ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=interaction(Class, Survived), conds = Age, group=1))
ggplot(data=titanic) + geom_mosaic(aes(weight=Freq, vars=Class, group=1, conds=Age, fill=Survived))

data(happy, package="productplots")
ggplot(data = happy) + geom_mosaic(aes(vars=c(happy), group=1))
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=c(happy), group=1))
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=c(health), fill=happy, group=1))
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=c(health), fill=happy, group=1), na.rm=TRUE)
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=interaction(health, sex, degree), fill=happy, group=1), na.rm=TRUE)
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=product(health, sex, degree), fill=happy, group=1), na.rm=TRUE)

# here is where a bit more control over the spacing of the bars would be helpful:
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=c(age), fill=happy, group=1), na.rm=TRUE)
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=c(age), fill=happy, conds = sex, group=1), na.rm=TRUE)
# facetting works!!!!
ggplot(data = happy) + geom_mosaic(aes(weight=wtssall, vars=c(age), fill=happy, group=1), na.rm=TRUE) + facet_grid(sex~.)

df <- read.csv("inst/mosaic-rects.csv")
ggplot() + geom_mosaic(aes(x = xmin, xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax), data=df)
}

