% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic-package.R
\docType{package}
\name{ggmosaic-package}
\alias{ggmosaic}
\alias{ggmosaic-package}
\title{ggmosaic: Mosaic Plots in the 'ggplot2' Framework}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Mosaic plots in the 'ggplot2' framework. Mosaic plot functionality is provided in a single 'ggplot2' layer by calling the geom 'mosaic'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://haleyjeppson.github.io/ggmosaic/}
  \item \url{https://github.com/haleyjeppson/ggmosaic}
  \item Report bugs at \url{https://github.com/haleyjeppson/ggmosaic}
}

}
\author{
\strong{Maintainer}: Haley Jeppson \email{hjeppson@iastate.edu} (\href{https://orcid.org/0000-0003-2524-4063}{ORCID})

Authors:
\itemize{
  \item Heike Hofmann \email{hofmann@iastate.edu} (\href{https://orcid.org/0000-0001-6216-5183}{ORCID})
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
}

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) [contributor]
}

}
\keyword{internal}
