% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.r
\name{margin}
\alias{margin}
\title{Internal helper function}
\usage{
margin(table, marginals = c(), conditionals = c())
}
\arguments{
\item{table}{data table}

\item{marginals}{variables used in margins of mosaic specification}

\item{conditionals}{variables conditioned upon in mosaic specification}
}
\value{
weighted data table
}
\description{
function copied directly from the productplots package
}
\author{
Hadley Wickham
}

