% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-mosaic.r
\name{product}
\alias{product}
\title{Wrapper for a list}
\usage{
product(...)
}
\arguments{
\item{...}{Unquoted variables going into the product plot.}
}
\description{
Wrapper for a list
}
\examples{
data(titanic)
ggplot(data=titanic) +
  geom_mosaic(aes(x=product(Survived, Class), fill=Survived))
}
