% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-mosaic.r
\name{product}
\alias{product}
\title{Product}
\usage{
product(x, ...)
}
\arguments{
\item{x}{variable}

\item{...}{other arguments passed on}
}
\description{
This function creates a product variable based on all combinations of the variables.
}

