% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\name{scale_product}
\alias{scale_product}
\alias{scale_x_product}
\alias{scale_y_product}
\title{Product scales for mosaic plots}
\usage{
scale_x_product(name = waiver(), breaks = waiver(),
  minor_breaks = waiver(), labels = waiver(), limits = NULL,
  expand = waiver(), oob = scales:::censor, na.value = NA_real_,
  trans = "identity")

scale_y_product(name = waiver(), breaks = waiver(),
  minor_breaks = waiver(), labels = waiver(), limits = NULL,
  expand = waiver(), oob = scales:::censor, na.value = NA_real_,
  trans = "identity")
}
\description{
product scales are especially introduced for use with mosaic plots: they are a hybrid of
continuous and discrete scales.
}

