% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\docType{data}
\name{scale_product}
\alias{ScaleContinuousProduct}
\alias{scale_product}
\alias{scale_x_product}
\alias{scale_y_product}
\title{Position scale for product variables}
\format{An object of class \code{ScaleContinuousProduct} (inherits from \code{ScaleContinuous}, \code{Scale}, \code{ggproto}) of length 2.}
\usage{
scale_x_product(name = waiver(), breaks = waiver(), labels = waiver(),
  minor_breaks = waiver(), limits = NULL, expand = waiver())

scale_y_product(name = waiver(), breaks = waiver(), labels = waiver(),
  minor_breaks = waiver(), limits = NULL, expand = waiver())

ScaleContinuousProduct
}
\arguments{
\item{name}{The name of the scale. Used as axis or legend title. If
\code{NULL}, the default, the name of the scale is taken from the first
mapping used for that aesthetic.}

\item{breaks}{One of: \itemize{
  \item \code{NULL} for no breaks
  \item \code{waiver()} for the default breaks computed by the
    transformation object
  \item A numeric vector of positions
  \item A function that takes the limits as input and returns breaks
    as output
}}

\item{labels}{One of: \itemize{
  \item \code{NULL} for no labels
  \item \code{waiver()} for the default labels computed by the
    transformation object
  \item A character vector giving labels (must be same length as \code{breaks})
  \item A function that takes the breaks as input and returns labels
    as output
}}

\item{minor_breaks}{One of: \itemize{
  \item \code{NULL} for no minor breaks
  \item \code{waiver()} for the default breaks (one minor break between
    each major break)
  \item A numeric vector of positions
  \item A function that given the limits returns a vector of minor breaks.
}}

\item{limits}{A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum.}

\item{expand}{A numeric vector of length two giving multiplicative and
additive expansion constants. These constants ensure that the data is
placed some distance away from the axes. The defaults are
\code{c(0.05, 0)} for continuous variables, and \code{c(0, 0.6)} for
discrete variables.}
}
\description{
Use \code{scale_*_product} with \code{product} variables.
Even though the breaks might suggest a discrete variable
it is probably easier to think of this scale as continuous.
}
\examples{
last_month <- Sys.Date() - 0:29
df <- data.frame(
  date = last_month,
  price = runif(30)
)
base <- ggplot(df, aes(date, price)) +
  geom_line()

}
\seealso{
\code{\link{scale_continuous}} for continuous position scales.
}
\keyword{datasets}

