% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\name{scale_type.productlist}
\alias{scale_type.productlist}
\title{Helper function that ggplot2 needs for determining scales on x and y}
\usage{
scale_type.productlist(x)
}
\arguments{
\item{x}{variable under consideration}
}
\value{
character string "productlist"
}
\description{
Helper function that ggplot2 needs for determining scales on x and y
}

