% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-product.R
\docType{data}
\name{scale_x_productlist}
\alias{scale_x_productlist}
\alias{scale_y_productlist}
\alias{ScaleContinuousProduct}
\title{Determining scales for mosaics}
\format{An object of class \code{ScaleContinuousProduct} (inherits from \code{ScaleContinuousPosition}, \code{ScaleContinuous}, \code{Scale}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
scale_x_productlist(name = waiver(), breaks = product_breaks(),
  minor_breaks = NULL, labels = product_labels(), limits = NULL,
  expand = waiver(), oob = scales:::censor, na.value = NA_real_,
  trans = "identity", position = "bottom", sec.axis = waiver())

scale_y_productlist(name = waiver(), breaks = product_breaks(),
  minor_breaks = NULL, labels = product_labels(), limits = NULL,
  expand = waiver(), oob = scales:::censor, na.value = NA_real_,
  trans = "identity", position = "left", sec.axis = waiver())

ScaleContinuousProduct
}
\arguments{
\item{sec.axis}{specify a secondary axis}
}
\description{
Determining scales for mosaics
}
\keyword{datasets}
